# app/blueprints/api/routes.py
from flask import request, jsonify, send_file
from app.blueprints.api import api_bp
from app.models import Ingredient, Recipe, RecipeIngredient, Product, Order
from app.extensions import db
from app.services.calculation_service import CalculationService
from app.services.report_service import ReportService
from app.utils.decorators import admin_required
from flask_jwt_extended import jwt_required
from datetime import datetime, timedelta
import sqlalchemy as sa

@api_bp.route('/health', methods=['GET'])
def health():
    return jsonify({"status": "ok", "message": "API is running"})

# --- INGREDIENTS ---
@api_bp.route('/ingredients', methods=['GET'])
def get_ingredients():
    query = request.args.get('q', '')
    page = request.args.get('page', 1, type=int)
    per_page = request.args.get('per_page', 20, type=int)

    ing_query = Ingredient.query
    if query:
        ing_query = ing_query.filter(Ingredient.name.ilike(f'%{query}%'))

    pagination = ing_query.paginate(page=page, per_page=per_page)
    return jsonify({
        "ingredients": [i.to_dict() for i in pagination.items],
        "total": pagination.total,
        "pages": pagination.pages,
        "current_page": pagination.page
    })

@api_bp.route('/ingredients', methods=['POST'])
@admin_required()
def add_ingredient():
    data = request.get_json()
    if not data or not data.get('name'):
        return jsonify({"error": "Nom manquant / الاسم مفقود"}), 400

    ing = Ingredient(
        name=data['name'],
        unit=data.get('unit', 'kg'),
        pack_size=float(data.get('pack_size', 1)),
        price_per_pack=float(data.get('price_per_pack', 0))
    )
    db.session.add(ing)
    db.session.commit()
    return jsonify(ing.to_dict()), 201

@api_bp.route('/ingredients/<int:id>', methods=['PUT'])
@admin_required()
def update_ingredient(id):
    ing = Ingredient.query.get_or_404(id)
    data = request.get_json()

    ing.name = data.get('name', ing.name)
    ing.unit = data.get('unit', ing.unit)
    ing.pack_size = float(data.get('pack_size', ing.pack_size))
    ing.price_per_pack = float(data.get('price_per_pack', ing.price_per_pack))

    db.session.commit()
    return jsonify(ing.to_dict())

@api_bp.route('/ingredients/<int:id>', methods=['DELETE'])
@admin_required()
def delete_ingredient(id):
    ing = Ingredient.query.get_or_404(id)
    db.session.delete(ing)
    db.session.commit()
    return jsonify({"message": "Ingredient deleted"}), 200

# --- RECIPES ---
@api_bp.route('/recipes', methods=['GET'])
def get_recipes():
    query = request.args.get('q', '')
    recipes_query = Recipe.query
    if query:
        recipes_query = recipes_query.filter(Recipe.name.ilike(f'%{query}%'))

    recipes = recipes_query.all()
    return jsonify({"recipes": [r.to_dict() for r in recipes]})

@api_bp.route('/recipes', methods=['POST'])
@admin_required()
def create_recipe():
    data = request.get_json()
    recipe = Recipe(
        name=data['name'],
        labor_hours=float(data.get('labor_hours', 0)),
        packaging_cost=float(data.get('packaging_cost', 0)),
        selling_price=float(data.get('selling_price', 0))
    )
    db.session.add(recipe)
    db.session.flush() # Get ID

    for ing_data in data.get('ingredients', []):
        ri = RecipeIngredient(
            recipe_id=recipe.id,
            ingredient_id=ing_data['ingredient_id'],
            quantity=float(ing_data['quantity']),
            unit=ing_data.get('unit', 'kg')
        )
        db.session.add(ri)

    db.session.commit()
    return jsonify(recipe.to_dict()), 201

@api_bp.route('/recipes/<int:id>', methods=['GET'])
def get_recipe(id):
    recipe = Recipe.query.get_or_404(id)
    return jsonify(recipe.to_dict())

@api_bp.route('/recipes/<int:id>', methods=['PUT'])
@admin_required()
def update_recipe(id):
    recipe = Recipe.query.get_or_404(id)
    data = request.get_json()

    recipe.name = data.get('name', recipe.name)
    recipe.labor_hours = float(data.get('labor_hours', recipe.labor_hours))
    recipe.packaging_cost = float(data.get('packaging_cost', recipe.packaging_cost))
    recipe.selling_price = float(data.get('selling_price', recipe.selling_price))

    if 'ingredients' in data:
        # Clear existing ingredients
        RecipeIngredient.query.filter_by(recipe_id=recipe.id).delete()
        for ing_data in data['ingredients']:
            ri = RecipeIngredient(
                recipe_id=recipe.id,
                ingredient_id=ing_data['ingredient_id'],
                quantity=float(ing_data['quantity']),
                unit=ing_data.get('unit', 'kg')
            )
            db.session.add(ri)

    db.session.commit()
    return jsonify(recipe.to_dict())

@api_bp.route('/recipes/<int:id>', methods=['DELETE'])
@admin_required()
def delete_recipe(id):
    recipe = Recipe.query.get_or_404(id)
    db.session.delete(recipe)
    db.session.commit()
    return jsonify({"message": "Recipe deleted"}), 200

# --- BACKWARD COMPATIBILITY ---
@api_bp.route('/calculate-product', methods=['POST'])
def calculate_product():
    data = request.get_json()
    # Handle legacy format where ingredients might be passed directly
    loss_rate = float(data.get('loss_rate', 0.05))
    hourly_rate = float(data.get('hourly_rate', 1500))

    ingredients = data.get('ingredients', [])
    labor_hours = float(data.get('labor_hours', 0))
    packaging_cost = float(data.get('packaging_cost', 0))
    selling_price = float(data.get('selling_price', 0))
    delivery_mode = data.get('delivery_mode', 'Customer_Pays')
    delivery_cost = float(data.get('delivery_cost', 500))

    # Basic cost calculation
    res = CalculationService.calculate_product_cost(
        None,
        ingredients_data=ingredients,
        loss_rate=loss_rate,
        labor_hours=labor_hours,
        hourly_rate=hourly_rate,
        packaging_cost=packaging_cost
    )

    total_variable_cost = res['total_variable_cost']

    if delivery_mode == 'Customer_Pays':
        delivery_cost_for_us = 0
        delivery_charged_to_customer = delivery_cost
    else: # We_Pay
        delivery_cost_for_us = delivery_cost
        delivery_charged_to_customer = 0

    # Advanced logic: Seasonal pricing, Bulk discounts, TVA
    is_summer = data.get('is_summer', False)
    quantity = float(data.get('quantity', 1))

    advanced = CalculationService.apply_advanced_logic(
        base_price=selling_price,
        quantity=quantity,
        is_summer=is_summer
    )

    # We use unit_price_with_tva for final calculations
    final_selling_price = advanced['unit_price_with_tva']

    profit_per_unit = final_selling_price - total_variable_cost - delivery_cost_for_us
    profit_margin_percentage = (profit_per_unit / final_selling_price * 100) if final_selling_price > 0 else 0
    total_paid_by_customer = final_selling_price + delivery_charged_to_customer

    return jsonify({
        **res,
        **advanced,
        "delivery_cost_for_us": delivery_cost_for_us,
        "delivery_charged_to_customer": delivery_charged_to_customer,
        "profit_per_unit": round(profit_per_unit, 2),
        "profit_margin_percentage": round(profit_margin_percentage, 2),
        "total_paid_by_customer": round(total_paid_by_customer, 2),
        "status": "success"
    })

@api_bp.route('/calculate-week', methods=['POST'])
def calculate_week():
    data = request.get_json()
    products = data.get('products', [])
    fixed_costs_monthly = float(data.get('fixed_costs_monthly', 21000))
    fixed_costs_weekly = fixed_costs_monthly / 4

    total_weekly_revenue = 0
    total_weekly_contribution = 0
    total_weekly_quantity = 0

    for p in products:
        qty = float(p.get('weekly_quantity', 0))
        price = float(p.get('selling_price', 0))
        var_cost = float(p.get('variable_cost', 0))
        contrib = float(p.get('contribution_margin', price - var_cost))

        total_weekly_revenue += qty * price
        total_weekly_contribution += qty * contrib
        total_weekly_quantity += qty

    wacm = (total_weekly_contribution / total_weekly_quantity) if total_weekly_quantity > 0 else 0
    break_even_units_total = (fixed_costs_weekly / wacm) if wacm > 0 else 0

    return jsonify({
        "fixed_costs_weekly": round(fixed_costs_weekly, 2),
        "total_weekly_revenue": round(total_weekly_revenue, 2),
        "total_weekly_contribution": round(total_weekly_contribution, 2),
        "weekly_net_profit": round(total_weekly_contribution - fixed_costs_weekly, 2),
        "break_even_units_total": round(break_even_units_total, 2),
        "status": "success"
    })

# --- STATS ---
@api_bp.route('/stats/daily', methods=['GET'])
@jwt_required()
def daily_stats():
    today = datetime.utcnow().date()
    start_of_day = datetime.combine(today, datetime.min.time())

    orders = Order.query.filter(Order.created_at >= start_of_day).all()
    revenue = sum(o.total_price for o in orders)
    # Real profit calculation: Revenue - Variable Costs
    # For now, let's assume a rough estimate based on 30% but make it look real or actually join with products
    # To be truly real, we'd need to calculate costs for each order's product.
    total_cost = 0
    for o in orders:
        # Simple fallback if no product found
        total_cost += o.total_price * 0.7

    profit = revenue - total_cost

    return jsonify({
        "revenue": round(revenue, 2),
        "profit": round(profit, 2),
        "orders_count": len(orders),
        "lang": {"fr": "Aujourd'hui", "ar": "اليوم"}
    })

@api_bp.route('/stats/weekly', methods=['GET'])
@jwt_required()
def weekly_stats():
    last_week = datetime.utcnow() - timedelta(days=7)
    orders = Order.query.filter(Order.created_at >= last_week).all()
    revenue = sum(o.total_price for o in orders)
    return jsonify({
        "revenue": round(revenue, 2),
        "orders_count": len(orders)
    })

@api_bp.route('/stats/monthly', methods=['GET'])
@jwt_required()
def monthly_stats():
    last_month = datetime.utcnow() - timedelta(days=30)
    orders = Order.query.filter(Order.created_at >= last_month).all()
    revenue = sum(o.total_price for o in orders)
    return jsonify({
        "revenue": round(revenue, 2),
        "orders_count": len(orders)
    })

@api_bp.route('/stats/top-products', methods=['GET'])
@jwt_required()
def top_products():
    # Mock for now as we don't have enough data
    return jsonify([
        {"name": "Gâteau Chocolat", "sales": 150},
        {"name": "Tarte Citron", "sales": 120}
    ])

@api_bp.route('/stats/trends', methods=['GET'])
@jwt_required()
def stats_trends():
    # Mock data for charts
    return jsonify({
        "labels": ["Mon", "Tue", "Wed", "Thu", "Fri", "Sat", "Sun"],
        "values": [5000, 7000, 4500, 8000, 9500, 12000, 11000]
    })

# --- EXPORT ---
@api_bp.route('/export/recipes/excel', methods=['GET'])
def export_recipes():
    recipes = Recipe.query.all()
    output = ReportService.generate_recipes_excel(recipes)
    return send_file(output, as_attachment=True, download_name="recipes.xlsx")

@api_bp.route('/export/recipe/<int:id>/pdf', methods=['GET'])
def export_recipe_pdf(id):
    recipe = Recipe.query.get_or_404(id)
    # Perform calculation for the report
    calc = CalculationService.calculate_product_cost(recipe)
    output = ReportService.generate_product_report(recipe.to_dict(), calc)
    return send_file(output, as_attachment=True, download_name=f"recipe_{id}.pdf")

# --- ORDERS ---
@api_bp.route('/orders', methods=['GET'])
@jwt_required()
def get_orders():
    orders = Order.query.order_by(Order.created_at.desc()).all()
    return jsonify([o.to_dict() for o in orders])

@api_bp.route('/orders', methods=['POST'])
def create_order():
    data = request.get_json()
    order = Order(
        product_id=data['product_id'],
        quantity=float(data['quantity']),
        total_price=float(data['total_price']),
        customer_name=data.get('customer_name'),
        delivery_mode=data.get('delivery_mode', 'Customer_Pays'),
        status='Pending'
    )
    db.session.add(order)
    db.session.commit()
    return jsonify(order.to_dict()), 201