# app/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Gestion Pâtisserie Alger API",
    "description": "API for bakery management in Algiers",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "basePath": "/api",
  "schemes": ["http"],
  "paths": {
    "/health": {
      "get": {
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/auth/register": {
      "post": {
        "parameters": [{ "name": "body", "in": "body", "schema": { "$ref": "#/definitions/User" } }],
        "responses": { "201": { "description": "User created" } }
      }
    },
    "/auth/login": {
      "post": {
        "parameters": [{ "name": "body", "in": "body", "schema": { "type": "object", "properties": { "username": { "type": "string" }, "password": { "type": "string" } } } }],
        "responses": { "200": { "description": "Login successful" } }
      }
    },
    "/recipes": {
      "get": {
        "responses": { "200": { "description": "List of recipes" } }
      },
      "post": {
        "parameters": [{ "name": "body", "in": "body", "schema": { "$ref": "#/definitions/Recipe" } }],
        "responses": { "201": { "description": "Recipe created" } }
      }
    },
    "/ingredients": {
      "get": {
        "responses": { "200": { "description": "List of ingredients" } }
      },
      "post": {
        "parameters": [{ "name": "body", "in": "body", "schema": { "$ref": "#/definitions/Ingredient" } }],
        "responses": { "201": { "description": "Ingredient created" } }
      }
    },
    "/stats/daily": {
      "get": {
        "responses": { "200": { "description": "Daily statistics" } }
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "username": { "type": "string" },
        "email": { "type": "string" },
        "password": { "type": "string" }
      }
    },
    "Recipe": {
      "type": "object",
      "properties": {
        "name": { "type": "string" },
        "selling_price": { "type": "number" }
      }
    },
    "Ingredient": {
      "type": "object",
      "properties": {
        "name": { "type": "string" },
        "price_per_pack": { "type": "number" },
        "pack_size": { "type": "number" }
      }
    }
  }
}